Internal rework for more (optionally) more control over individual rendering aspects moving forward.
Each shader can now have multiple pipeline configurations associated with it that can be swapped in and out. These are handled/configured with the 'ShaderProfile' class which takes some of the load off the ShaderLayout class to handle more configuration/pipeline related tasks.

Removed the ShaderPostProcessor and ShaderPreProcessor classes from the main shader class, these are now constructed on the fly and used exclusively for when the sources or binaries need to be loaded in again.

Removed staging buffer from the vertex and index buffers for the time being, they likely need to be split up into static and dynamic versions in the future.

Device initialization improvements to support multiple queues with preferences for dedicated queue families for things like transfer/compute queues.

Framework for dynamic pipeline states so that upon being bound the correct commands will always be recorded automatically with whatever values are currently set for the profile.

Removed the redundant 'Pipeline' and 'Pipeline Layout' classes

General improvements/optimizations to the handling of lower level vulkan objects primarily in regards to the deletion queue and debug messenger.

RenderPass overhaul to allow for more flexible/easy render pass system. This goes hand in hand with the new Shader Profile system and includes new methods to build render passes, and a new render pass scheduler.

Reduced the dependence that RenderPass and Shaders have with each other so that they both don't HAVE to be aware of each other's existence.

Overhauled the way command recording is handled to make it simpler, faster, easier to work with and more flexible. Now, only a single command buffer is needed for all render passes. The RenderCommand class has been altered to be more of a lower level functioning class and a new CommandPool class has replaced the old CommandBuffer class with a clearer purpose and job, greatly simplifying the whole process.

Synchronization with fences/semaphores has been greatly improved, proper fence implementation is now used to ensure swapchain images are fully acquired before being used, command buffer's synchronization between different frames in flight has been optimized and made much more reliable, safe and clear to understand as well as the synchronization of queues and present info using semaphores.

Various other more minor changes and fixes have been applied all over to make the whole engine much more reliable and less janky/hacky to work with.