Overhaul of the way render passes are handled. There is now a RenderChain class which consists of a collection of render passes that are executed in order depending on which are active and which aren't. Then the chain scheduler ties into the global renderer class to handle the execution of different chains/passes.
Render passes now coincide with a new RenderTarget class that replaces the previous framebuffer mechanism with a much more scalable solution.

There is no longer an image, imageview, and framebuffer (at least) per frame in flight for the selection and viewport passes. Instead, there is now a viewport render target and a selection viewport target, neither of which require having resources that persist accross frames.

While I can't be 100% certain yet, there is a very noticable speed improvement here (around 300-320 fps compared to the 240 or so it was running at before) which I assume for now is due to not having wasteful duplicate resources per frame in flight.

It needs more cleanup and improvement to be fully usable but as of now some very solid groundwork is laid out so improvements and additions should be much easier to implement.