Removed/cleaned up alot of nonsense, particularly with the gum library. We're experimenting with a flipped viewport in vulkan now to simulate a y up coordinate system and hopefully reduce alot of the issues that have been popping up relating to vertex/index values.
Most the work done in this commit actually involved reducing complexity with gum instead of adding complexity. Padding, margins, layouts, transforms etc are currently out, instead we're gonna start simple with ensuring the coordinate systems, bounds detection, relationships between components and rendering are all solid before adding anything fancy.

As a side note, it's looking pretty clear that Gum should function as a standalone library, so it'd be a good idea to start looking for areas it can be 'untangled' from the rest of Candy. The Math library can obviously be seperated and added as a dependency fairly easily. In it's current state it probably wouldn't be too hard to scrap it's usage of the graphics library since most of those abstractions are for general purpose uses and Gum will be much more specialized, but we'll have to consider possible advantages of keeping the two together.