Layouts for ui elements, correct collision detection for ui elements with rounded corners, min/pref/max sizing hints, layout calculations are now done in post order traversal so layouts/containers have the opportunity to resize themselves based on their children's layouts, optional spacing between elements, region element allows for custom shapes to define its collision detection and shape while still serving as a generic container element.
Minor additions to the Vector2 class and Bounds2D class.

Began rewriting the event system for gum to not depend on the one used in CandyEngine (eventually need to replace CandyEngine's with a slightly modified version of Gum, or make a generic event library and just add custom functionality for Gum and Candy Engine seperately).
The new event system plans are far more robust and flexible and (in theory) should massively speed up workflow when dealing with events. I'm debating a multithreaded event system, but especially for the UI side of things I'm not sure it would be worth it at the moment.

Note:
Eventually the Vector/Math classes need some added functionality like Abs, Max, Min, etc to be closer in line with the way it works in glsl.